+++ mod/evaluate.lua
@@ -263,14 +263,20 @@ end
 ---@param expression string
 ---@param seq number
 function __DebugAdapter.evaluate(frameId,context,expression,seq)
-  if not frameId then
-    -- if you manage to do one of these fast enough for data, go for it...
-    if not data and __DebugAdapter.canRemoteCall() and script.mod_name~="level" then
-      -- remote to `level` if possible, else just error
-      if remote.interfaces["__debugadapter_level"] then
-        return remote.call("__debugadapter_level","evaluate",frameId,context,expression,seq)
+  -- if you manage to do one of these fast enough for data, go for it...
+  if not frameId and not data then
+    local modname,rest = expression:match("^__(.-)__ (.+)$")
+    if modname then
+      expression = rest
+    else
+      modname = "level"
+    end
+    if script.mod_name~=modname then
+      -- remote to named state if possible, else just error
+      if __DebugAdapter.canRemoteCall() and remote.interfaces["__debugadapter_"..modname] then
+        return remote.call("__debugadapter_"..modname,"evaluate",frameId,context,expression,seq)
       else
-        print("DBGeval: " .. json.encode({result = "`level` not available for eval", type="error", variablesReference=0, seq=seq}))
+        print("DBGeval: " .. json.encode({result = "`"..modname.."` not available for eval", type="error", variablesReference=0, seq=seq}))
         return
       end
     end